+++ src/main/java/ru/romangr/catbot/telegram/model/DocumentMessageToSend.kt
@@ -3,6 +3,6 @@ package ru.romangr.catbot.telegram.model
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class DocumentMessageToSend(
-        @JsonProperty("chat_id") val chatId: Int,
+        @JsonProperty("chat_id") val chatId: Long,
         @JsonProperty("document") val documentId: String
 )
